Test fixes - part 1

+++ crates/jmap/src/api/management/principal.rs
@@ -13,7 +13,7 @@ use directory::{
         manage::{self, not_found, ManageDirectory},
         PrincipalAction, PrincipalField, PrincipalUpdate, PrincipalValue, SpecialSecrets,
     },
-    DirectoryInner, Permission, Principal, QueryBy, Type, ROLE_ADMIN, ROLE_TENANT_ADMIN, ROLE_USER,
+    DirectoryInner, Permission, Principal, QueryBy, Type,
 };
 
 use hyper::{header, Method};
@@ -89,54 +89,6 @@ impl JMAP {
                     self.assert_supported_directory()?;
                 }
 
-                // Validate tenant limits
-                #[cfg(feature = "enterprise")]
-                if self.core.is_enterprise_edition() {
-                    if let Some(tenant_info) = access_token.tenant {
-                        let tenant = self
-                            .core
-                            .storage
-                            .data
-                            .query(QueryBy::Id(tenant_info.id), false)
-                            .await?
-                            .ok_or_else(|| {
-                                trc::ManageEvent::NotFound
-                                    .into_err()
-                                    .caused_by(trc::location!())
-                            })?;
-
-                        // Enforce tenant quotas
-                        if let Some(limit) = tenant
-                            .get_int_array(PrincipalField::Quota)
-                            .and_then(|quotas| quotas.get(principal.typ() as usize + 1))
-                            .copied()
-                            .filter(|q| *q > 0)
-                        {
-                            // Obtain number of principals
-                            let total = self
-                                .core
-                                .storage
-                                .data
-                                .count_principals(
-                                    None,
-                                    principal.typ().into(),
-                                    tenant_info.id.into(),
-                                )
-                                .await
-                                .caused_by(trc::location!())?;
-
-                            if total >= limit {
-                                trc::bail!(trc::LimitEvent::TenantQuota
-                                    .into_err()
-                                    .details("Tenant principal quota exceeded")
-                                    .ctx(trc::Key::Details, principal.typ().as_str())
-                                    .ctx(trc::Key::Limit, limit)
-                                    .ctx(trc::Key::Total, total));
-                            }
-                        }
-                    }
-                }
-
                 // Create principal
                 let result = self
                     .core
@@ -154,20 +106,59 @@ impl JMAP {
                 // List principal ids
                 let params = UrlParams::new(req.uri().query());
                 let filter = params.get("filter");
-                let typ = params.parse("type").unwrap_or(Type::Individual);
                 let page: usize = params.parse("page").unwrap_or(0);
                 let limit: usize = params.parse("limit").unwrap_or(0);
 
+                // Parse types
+                let mut types = Vec::new();
+                for typ in params
+                    .get("types")
+                    .or_else(|| params.get("type"))
+                    .unwrap_or_default()
+                    .split(',')
+                {
+                    if let Some(typ) = Type::parse(typ) {
+                        if !types.contains(&typ) {
+                            types.push(typ);
+                        }
+                    }
+                }
+
+                // Parse fields
+                let mut fields = Vec::new();
+                for field in params.get("fields").unwrap_or_default().split(',') {
+                    if let Some(field) = PrincipalField::try_parse(field) {
+                        if !fields.contains(&field) {
+                            fields.push(field);
+                        }
+                    }
+                }
+
                 // Validate the access token
-                access_token.assert_has_permission(match typ {
-                    Type::Individual => Permission::IndividualList,
-                    Type::Group => Permission::GroupList,
-                    Type::List => Permission::MailingListList,
-                    Type::Domain => Permission::DomainList,
-                    Type::Tenant => Permission::TenantList,
-                    Type::Role => Permission::RoleList,
-                    Type::Resource | Type::Location | Type::Other => Permission::PrincipalList,
-                })?;
+                let validate_types = if !types.is_empty() {
+                    types.as_slice()
+                } else {
+                    &[
+                        Type::Individual,
+                        Type::Group,
+                        Type::List,
+                        Type::Domain,
+                        Type::Tenant,
+                        Type::Role,
+                        Type::Other,
+                    ]
+                };
+                for typ in validate_types {
+                    access_token.assert_has_permission(match typ {
+                        Type::Individual => Permission::IndividualList,
+                        Type::Group => Permission::GroupList,
+                        Type::List => Permission::MailingListList,
+                        Type::Domain => Permission::DomainList,
+                        Type::Tenant => Permission::TenantList,
+                        Type::Role => Permission::RoleList,
+                        Type::Resource | Type::Location | Type::Other => Permission::PrincipalList,
+                    })?;
+                }
 
                 let mut tenant = access_token.tenant.map(|t| t.id);
 
@@ -186,32 +177,19 @@ impl JMAP {
                                 .map(|p| p.id);
                         }
                     }
-                } else if matches!(typ, Type::Tenant) {
+                } else if types.contains(&Type::Tenant) {
                     return Err(manage::enterprise());
                 }
 
-                let accounts = self
+                let principals = self
                     .core
                     .storage
                     .data
-                    .list_principals(filter, typ.into(), tenant)
+                    .list_principals(filter, tenant, &types, &fields, page, limit)
                     .await?;
 
-                let (total, accounts) = if limit > 0 {
-                    let offset = page.saturating_sub(1) * limit;
-                    (
-                        accounts.len(),
-                        accounts.into_iter().skip(offset).take(limit).collect(),
-                    )
-                } else {
-                    (accounts.len(), accounts)
-                };
-
                 Ok(JsonResponse::new(json!({
-                        "data": {
-                            "items": accounts,
-                            "total": total,
-                        },
+                        "data": principals,
                 }))
                 .into_http_response())
             }
@@ -256,64 +234,13 @@ impl JMAP {
                             .await?
                             .ok_or_else(|| trc::ManageEvent::NotFound.into_err())?;
 
-                        // Map groups
-                        for field in [
-                            PrincipalField::MemberOf,
-                            PrincipalField::Lists,
-                            PrincipalField::Roles,
-                        ] {
-                            if let Some(member_of) = principal.take_int_array(field) {
-                                for principal_id in member_of {
-                                    match principal_id as u32 {
-                                        ROLE_ADMIN if field == PrincipalField::Roles => {
-                                            principal.append_str(field, "admin");
-                                        }
-                                        ROLE_TENANT_ADMIN if field == PrincipalField::Roles => {
-                                            principal.append_str(field, "tenant-admin");
-                                        }
-                                        ROLE_USER if field == PrincipalField::Roles => {
-                                            principal.append_str(field, "user");
-                                        }
-                                        principal_id => {
-                                            if let Some(name) = self
-                                                .core
-                                                .storage
-                                                .data
-                                                .get_principal_name(principal_id)
-                                                .await
-                                                .caused_by(trc::location!())?
-                                            {
-                                                principal.append_str(field, name);
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // Obtain quota usage
-                        if matches!(typ, Type::Individual | Type::Group | Type::Tenant) {
-                            principal.set(
-                                PrincipalField::UsedQuota,
-                                self.get_used_quota(account_id).await? as u64,
-                            );
-                        }
-
-                        // Obtain member names
-                        for member_id in self.core.storage.data.get_members(account_id).await? {
-                            if let Some(mut member_principal) = self
-                                .core
-                                .storage
-                                .data
-                                .query(QueryBy::Id(member_id), false)
-                                .await?
-                            {
-                                if let Some(name) = member_principal.take_str(PrincipalField::Name)
-                                {
-                                    principal.append_str(PrincipalField::Members, name);
-                                }
-                            }
-                        }
+                        // Map fields
+                        self.core
+                            .storage
+                            .data
+                            .map_field_ids(&mut principal, &[])
+                            .await
+                            .caused_by(trc::location!())?;
 
                         Ok(JsonResponse::new(json!({
                                 "data": principal,
@@ -334,15 +261,17 @@ impl JMAP {
                             }
                         })?;
 
-                        // Remove FTS index
-                        self.core.storage.fts.remove_all(account_id).await?;
-
                         // Delete account
                         self.core
                             .storage
                             .data
                             .delete_principal(QueryBy::Id(account_id))
                             .await?;
+
+                        // Remove FTS index
+                        if matches!(typ, Type::Individual | Type::Group) {
+                            self.core.storage.fts.remove_all(account_id).await?;
+                        }
 
                         // Remove entries from cache
                         self.inner.sessions.retain(|_, id| id.item != account_id);
